remove person-sharing in favor of colocation
